refactor: move helper function to url utils
